Main script
This script allows the user to upload their items by scanning barcodes in the inventory. The inventory will integrate the items in a database which will hold all the information regarding purchase and expiration date of specific items and will notify the user by sending an alert based on user's settings.